Fixed undefined-behaviour in Endian header
For some reason, converting to a 'byte' pointer instead of a 'char'
pointer was causing the optimizer to not emit code for swapping the
endianness of floats/doubles. This may be an issue with it treating
it like a violation of strict-aliasing; however, converting to a
pointer that is related to a 'char' (such as an enum whose base is
a char) should not violate strict-aliasing. This requires investigation,
since the majority of the library uses the 'byte' type for pointer
adjustments.