Removed Allocator::is_always_equal optional concept
Standard allocators make use of `is_always_equal` to determine whether
an allocator from one container can be used to deallocate the memory
from another container.

Since allocators in this library do not contain the underlying type
being allocated as part of the type signature, such information is
always limited to whether or not an allocation was allocated from the
same instance as the container. The only time where `is_always_equal`
would safely be `std::true_type` is when an allocator is stateless,
since a deallocation does not require tracking of state. Thus,
`is_always_equal` is now removed and should be determined/implied by
`is_stateless`