Add unit tests for block allocators (#28)
All block allocators now have passing unit tests that attempt
some level of code-coverage. Additionally, unit tests also 
encompass static validation of types to ensure that 
each allocator accurately passes the `is_block_allocator` 
type trait/concept.

This will not be fully comprehensive, due to the number of 
possible instantiations and alternative code-paths that can
be taken with templates -- but it will cover the general cases
along with syntactical validation (which ensures compilers like
MSVC treat the SFINAE correctly).

A few changes to some of the block allocators were necessary 
in order for tests to pass. These changes were mostly minimal,
and should not affect the public-facing API in a major way.