Add storage concepts (#29)
The previous library implementation was lacking in the 
overall concepts. `Allocator` and `BlockAllocator` types 
were not required to be copyable/moveable (and neither
should they be, since they may manage local resources).
This makes composition complicated, since an `Allocator`
may have to own (and possibly EBO-inherit) from a `BlockAllocator`
to extend its capability. 

As a manner around that, new concepts have been added for
`AllocatorStorage` and `BlockAllocatorStorage`. These types
simply have `get_allocator()` and `get_block_allocator()` functions
respectively, with `allocator_type` and `block_allocator_type` member
types. These are required to be non-throwing copyable/moveable to
allow for a simpler way of storing types. As an additional benefit, this
also means that the storage policies can also manage the lifetime
-- which can allow for stateful `Allocator` or `BlockAllocator` types to
be referenced instead of strongly-owned. 

Removed:
- `allocator_reference` 
- `min_aligned_allocator`

Added:
- new concepts for `AllocatorStorage` and `BlockAllocatorStorage`
- `referenced_allocator_storage`
- `shared_allocator_storage`
- `stateless_allocator_storage`
- `referenced_block_allocator_storage`
- `shared_block_allocator_storage`
- `stateless_block_allocator_storage`

Renamed:
- ` any_allocator` -> `allocator_reference`
- `any_block_allocator` -> `block_allocator_reference`

Updated:
- Composite allocators to mostly use `AllocatorStorage`