Invoke 'make'/'dispose' from allocator
allocator_traits will now delegate to the underlying allocator first
for the following functions:
* try_make
* try_make_array
* make
* make_array
* dispose
* dispose_array

In the event that one of these functions is not defined, a custom
implementation is instead invoked that performs these functionalities
out of 'allocate' and 'deallocate' calls.

Additionally, these functions have been written to support
'pretty pointers', so that custom pointer types can be returned. This
is done through rebinding 'pointer' to a typed pointer instead.