Invoke 'make'/'dispose' from allocator in allocator_traits (#30)
allocator_traits will now delegate to the underlying allocator first
for the following functions:
* make
* make_array
* dispose
* dispose_array

In the event that one of these functions is not defined, a custom
implementation is instead invoked that performs these functionalities
out of 'allocate' and 'deallocate' calls.

Additionally, these functions have been written to support
'pretty pointers', so that custom pointer types can be returned. This
is done through rebinding 'pointer' to a typed pointer instead.

Unfortunately, MSVC had some issues with the alias `typed_pointer`, 
and the implementations of these had to be written with fully
expanded types, rather than nice clean aliases.