Restructure files into subsystems (#33)
* Rename 'aligned_memory' -> 'aligned_heap_memory'

The aligned memory file has been updated to include 'heap' in the name,
so it doesn't get confused with other terms that may use 'memory' in
the name.

* Rename 'unaligned_memory' -> 'unaligned_storage'

This change is to disambiguate headers that contain memory regions,
and headers that contain utility functions

* Move memory allocating utilities to 'regions'

This adds a new subdirectory 'regions' to the library that contains the
miscellaneous memory regions used for allocations in this library

* Move top-level headers to 'utilities'

All remaining top-level headers have been moved to the subdirectory
'utilities'

* Move 'standard_allocator' to 'adapters'

* Move trait classes to 'traits'

allocator_traits and block_allocator_traits were moved to a
new subdirectory 'traits' to remove a cyclic dependency.

Prior to this, utilities and concepts had a cyclic dependency wherein
concepts were made out of utility types, and [block_]allocator_traits
had a dependency to concepts. With this move, there is no more
intersubsystem cyclic dependency.

* Move deleters to 'adapters'

The deleter types fundamentally exist purely for interop with the STL's
std::shared_ptr and std::unique_ptr types. Though the concept exists
in this library, it's primary intention is to adapt to work with the
STL -- and thus this is a better place for it.

* Promote dynamic_size_type to upper namespace

This removes the dynamic_size_type from the detail namespace, and
promotes it to be in the bit::memory namespace instead for better
visibility

* Promote ebo_storage to upper namespace

ebo_storage has been removed from the 'detail' namespace and promoted
to the bit::memory namespace for better visibility

* Move policies to 'policies'

Taggers, trackers, bounds checkers, and lockables have all been
relocated to a 'policies' directory for use with the
policy-based allocators.