Remove equality requirement for allocators
The requirement for allocators and block_allocators to have
equality operators has been lifted. For 2 allocators to ever
be considered equal, they must both manage the same underlying
memory, at the same state. Since every allocator uniquely owns
the memory in which it points to, the only way for this equality
to ever be true is for the allocator to be comparing against itself.