Extend Allocator Traits (#39)
* Add address type
This adds a new 'address' integral type. This is effectively a type-safe
version of std::uintptr_t that prevents the possibility of casting a
const T* to an integral value and back to a T*. The address and
const_address types both work to allow performing integral
operations on an otherwise type-safe integral representation of
a pointer's address

* Add BIT_MEMORY_UNREACHABLE
Certain codepaths should be considered completely unreachable, and
should be documented with this in order to promote better compiler
optimizations.
This is used in allocator_traits to instruct the compiler that invocation
of an out-of-memory handler must terminate the current codepath.

* Optimize 'destroy_at' for pointer_utilities
'destroy_at' will not call a pseudo-destructor if the type is trivial

* Extend allocator_traits
allocator_traits has been reworked to now include pointer conversions
from raw to pretty pointers and back.
Additionally, the extended-only logic of allocator_traits has now been
broken into a new traits type of 'extended_allocator_traits'. This
removes the need for SFINAE on non-existent implementations of
'try_allocate' with the extended offset.